@@ -4,23 +4,31 @@
   "description": "Nashville Journal",
   "author": "Davis Hunt",
   "dependencies": {
+    "babel": "^6.5.2",
     "babel-core": "^5.8.24",
     "babel-loader": "^5.3.2",
     "classnames": "^2.2.3",
+    "intui": "git+https://github.com/arxii/intui.git",
     "jquery": "^2.1.3",
     "moment": "^2.9.0",
     "moment-timezone": "^0.4.1",
-    "react": "^0.13.3",
+    "react": "^0.14.8",
+    "react-dom": "^0.14.8",
     "whatwg-fetch": "^0.11.0"
   },
   "devDependencies": {
+    "babel-core": "^5.8.38",
+    "babel-loader": "^5.4.0",
+    "babel-preset-es2015": "^6.6.0",
     "babelify": "^6.3.0",
     "grunt": "^0.4.5",
+    "grunt-autoprefixer": "^3.0.3",
     "grunt-browserify": "^3.4.0",
     "grunt-contrib-copy": "^0.8.0",
     "grunt-contrib-sass": "^0.9.2",
     "grunt-contrib-uglify": "^0.8.0",
     "grunt-contrib-watch": "^0.6.1",
-    "grunt-autoprefixer": "^3.0.3"
+    "gulp": "^3.9.1",
+    "gulp-sass": "^2.2.0"
   }
 }
